Move minimum commission decorator to its own file
