+++ module/app/ante/ante.go
@@ -5,9 +5,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	sdkante "github.com/cosmos/cosmos-sdk/x/auth/ante"
-	"github.com/cosmos/cosmos-sdk/x/authz"
-	"github.com/cosmos/cosmos-sdk/x/bank/keeper"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	ibcante "github.com/cosmos/ibc-go/v3/modules/core/ante"
 	ibckeeper "github.com/cosmos/ibc-go/v3/modules/core/keeper"
 )
@@ -54,80 +51,4 @@ func chainHandlers(handler sdk.AnteHandler, secondHandler sdk.AnteHandler) sdk.A
 		// Next run the second handler
 		return secondHandler(newCtx, tx, simulate)
 	}
-}
-
-// MinCommissionDecorator was originally included in the Juno project at github.com/CosmosContracts/juno/app/ante.go
-// This version was added to Juno by github user the-frey https://github.com/the-frey and Giancarlos Salas https://github.com/giansalex
-// The Juno project obtained their initial version of this decorator from github.com/public-awesome/stargaze with original
-// author Jorge Hernandez https://github.com/jhernandezb
-type MinCommissionDecorator struct {
-	cdc codec.BinaryCodec
-}
-
-func NewMinCommissionDecorator(cdc codec.BinaryCodec) MinCommissionDecorator {
-	return MinCommissionDecorator{cdc}
-}
-
-func (min MinCommissionDecorator) AnteHandle(
-	ctx sdk.Context, tx sdk.Tx,
-	simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
-	msgs := tx.GetMsgs()
-	minCommissionRate := sdk.NewDecWithPrec(5, 2)
-
-	validMsg := func(m sdk.Msg) error {
-		switch msg := m.(type) {
-		case *stakingtypes.MsgCreateValidator:
-			// prevent new validators joining the set with
-			// commission set below 5%
-			c := msg.Commission
-			if c.Rate.LT(minCommissionRate) {
-				return sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "commission can't be lower than 5%")
-			}
-		case *stakingtypes.MsgEditValidator:
-			// if commission rate is nil, it means only
-			// other fields are affected - skip
-			if msg.CommissionRate == nil {
-				break
-			}
-			if msg.CommissionRate.LT(minCommissionRate) {
-				return sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "commission can't be lower than 5%")
-			}
-		}
-
-		return nil
-	}
-
-	validAuthz := func(execMsg *authz.MsgExec) error {
-		for _, v := range execMsg.Msgs {
-			var innerMsg sdk.Msg
-			err := min.cdc.UnpackAny(v, &innerMsg)
-			if err != nil {
-				return sdkerrors.Wrapf(sdkerrors.ErrUnauthorized, "cannot unmarshal authz exec msgs")
-			}
-
-			err = validMsg(innerMsg)
-			if err != nil {
-				return err
-			}
-		}
-
-		return nil
-	}
-
-	for _, m := range msgs {
-		if msg, ok := m.(*authz.MsgExec); ok {
-			if err := validAuthz(msg); err != nil {
-				return ctx, err
-			}
-			continue
-		}
-
-		// validate normal msgs
-		err = validMsg(m)
-		if err != nil {
-			return ctx, err
-		}
-	}
-
-	return next(ctx, tx, simulate)
 }
